Moved result into its own file
